requirements.txt
# Iron Arena Backend Dependencies
# Created by: engineering-backend-architect
# Updated by: engineering-backend-architect (added PyJWT, apscheduler, python-dotenv)
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
python-telegram-bot==21.1.1
pydantic==2.7.1
pydantic-settings==2.2.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
websockets==12.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.1
